Fix IP rate limiter
